Add basic brute force for everything except plugs
This is mainly to see how quickly we can brute force the rotor choice,
order, start positions and 2 of 3 ring settings. This assumes we know
the plugboard (which is the hardest part of cracking the Enigma, so we
are very much cheating).

This will still be useful though, as once we implement the actual Bombe
machine, brute forcing the rest of the settings and remaining plug pairs
will still need to happen.